+++ Editor/Scripts/TMP3D_UnlitShaderGUI.cs
@@ -35,8 +35,8 @@ namespace Ikaroon.TMP3DEditor
 			{
 				undoLabel = "RaymarchMode",
 				label = new GUIContent("Raymarch Mode"),
-				keywords = new[] { "_RAYMARCHER_STANDARD", "_RAYMARCHER_PENALTY" },
-				keywordLabels = new[] { new GUIContent("Standard"), new GUIContent("Penalty") }
+				keywords = new[] { "_RAYMARCHER_SDF", "_RAYMARCHER_SIMPLE" },
+				keywordLabels = new[] { new GUIContent("Signed Distance Field"), new GUIContent("Simple") }
 			};
 			s_maxStepsFeature = new TMP3D_ShaderFeature()
 			{
@@ -123,7 +123,15 @@ namespace Ikaroon.TMP3DEditor
 			s_raymarchModeFeature.DoPopup(m_Editor, m_Material);
 			s_maxStepsFeature.DoPopup(m_Editor, m_Material);
 
-			DoSlider("_RaymarchMinStep", "Min Step");
+			switch (s_raymarchModeFeature.State)
+			{
+				case 0:
+					DoSlider("_RaymarchMinStep", "Min Step");
+					break;
+				case 1:
+					DoSlider("_RaymarchStepLength", "Step Length");
+					break;
+			}
 			DoTexture2D("_DepthAlbedo", "Depth Albedo");
 
 			EditorGUI.indentLevel -= 1;